Create Motor_control_Joystick_90Degrees
In this code we control a stepper motor with a Joystick and an ADC module "MCP3426" to move 90 degrees left and right, in syncron with the value from the joystick. For example, if we move a little from the joystick to left, the motor will meve slightly to left, but if we move the joystick to it's maximum value, the motor will go the the maximum value of 90 dergees.