Create App_connect
This code is creating an app to login with the username/password and to redirect us to another application or folder. We used MariaDB server and database to store the username and passsword of each user, and encrypted the password. For the code to hash the password see next script.